requirements-celery.txt
Django==5.0.13
celery==5.3.6
redis==4.6.0
flower==2.0.1
kombu==5.3.6
django-celery-results>=2.5,<3.0
billiard>=4.1,<5.0
pymongo==4.11.3
pyodbc==5.2.0
python-dotenv==1.1.0
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.4
soupsieve==2.6